Fix missing bcm2710-rpi-3-b-plus.dtb
On relase 2018.08 Buildroot has included support to rpi-3-b-plus.
This include a new dtb target "bcm2710-rpi-3-b-plus" which isn't
compiled on olders Kernel version.

Even if the target board is not a rpi-3-b-plus, using the 2018.08
release would not work because of post-image script that tries to
copy the device tree blob to generate the sdcard image.

This commit update the Kernel of raspberrypi3_wpe_rdk_defconfig to
4.14.